Update and rename aa.js to a.js

+++ a.js
@@ -2,7 +2,7 @@ let id = localStorage.playerId
 let xor = id.charCodeAt(id.length - 337 % id.length - 1) ^ 82
 let uint8 = new Uint8Array([4, ...new TextEncoder().encode("Hello!")])
 
-for(let i = 0; i < uint.length; i++) uint[i] ^= xor
+for(let i = 0; i < uint8.length; i++) uint8[i] ^= xor
 [ uint8[0], uint8[337%uint8.length] ] = [ uint8[337%uint8.length], uint8[0] ]
 
 console.log(uint8, xor, id)
